Fix contextual replies.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@51111 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ elmdentica/src/elmdentica.h
@@ -34,7 +34,7 @@ typedef struct _ub_bubble {
 	char *message;
 	int   account_id;
 	int   account_type;
-	long int   status_id;
+	long long int   status_id;
 	Eina_Bool	favorite;
 } ub_Bubble;
